(BSR)[API] fix: remove repository/clean_all_database.py
