+++ api/src/pcapi/db_utils.py
@@ -2,6 +2,7 @@ import typing
 
 import sqlalchemy as sa
 from sqlalchemy import exc as sa_exc
+from sqlalchemy import text
 
 import pcapi.core.bookings.models as bookings_models
 import pcapi.core.criteria.models as criteria_models
@@ -28,6 +29,39 @@ from pcapi.models.beneficiary_import import BeneficiaryImport
 from pcapi.models.beneficiary_import_status import BeneficiaryImportStatus
 from pcapi.models.feature import Feature
 from pcapi.models.feature import install_feature_flags
+
+
+def install_database_extensions() -> None:
+    _create_text_search_configuration_if_not_exists()
+    _create_index_btree_gist_extension()
+    _create_postgis_extension()
+
+
+def _create_text_search_configuration_if_not_exists() -> None:
+    with db.engine.begin() as connection:
+        connection.execute(text("CREATE EXTENSION IF NOT EXISTS unaccent;"))
+
+        french_unaccent_configuration_query = connection.execute(
+            text("SELECT * FROM pg_ts_config WHERE cfgname='french_unaccent';")
+        )
+        if french_unaccent_configuration_query.fetchone() is None:
+            connection.execute(text("CREATE TEXT SEARCH CONFIGURATION french_unaccent ( COPY = french );"))
+            connection.execute(
+                text(
+                    "ALTER TEXT SEARCH CONFIGURATION french_unaccent"
+                    " ALTER MAPPING FOR hword, hword_part, word WITH unaccent, french_stem;"
+                )
+            )
+
+
+def _create_index_btree_gist_extension() -> None:
+    with db.engine.begin() as connection:
+        connection.execute(text("CREATE EXTENSION IF NOT EXISTS btree_gist;"))
+
+
+def _create_postgis_extension() -> None:
+    with db.engine.begin() as connection:
+        connection.execute(text("CREATE EXTENSION IF NOT EXISTS postgis;"))
 
 
 # Order of table to clean matters because of foreign key constraints
